+++ CHANGES.md
@@ -54,6 +54,9 @@ defined in *[@planetarium/account]*.
  -  Changed `BlockCommit` to implement `IBencodable`.  [[#2955]]
      -  Changed `BlockCommit(Dictionary)` to `BlockCommit(IValue)`.
      -  Changed `BlockCommit.Encoded` to `BlockCommit.Bencoded`.
+ -  Changed `BlockHash` to implement `IBencodable`.  [[#2959]]
+     -  Changed `BlockHash(Binary)` to `BlockHash(IValue)`.
+     -  Added `BlockHash.Bencoded` property.
 
 ### Backward-incompatible network protocol changes
 
@@ -85,6 +88,7 @@ defined in *[@planetarium/account]*.
 [#2915]: https://github.com/planetarium/libplanet/pull/2915
 [#2954]: https://github.com/planetarium/libplanet/pull/2954
 [#2955]: https://github.com/planetarium/libplanet/pull/2955
+[#2959]: https://github.com/planetarium/libplanet/pull/2959
 [@planetarium/account-web3-secret-storage]: https://www.npmjs.com/package/@planetarium/account-web3-secret-storage
 [@planetarium/bencodex]: https://www.npmjs.com/package/@planetarium/bencodex
 [disjukr/bencodex]: https://github.com/disjukr/bencodex