+++ CHANGES.md
@@ -6,6 +6,13 @@ Version 0.51.3
 
 To be released.
 
+### Bug fixes
+
+ -  (Libplanet.Net) Fixed a bug where `DealerSocket` cannot connect to peers
+    because of IPv4 return of `ResolveNetMQAddressAsync()`.  [[#2956]]
+
+[#2956]: https://github.com/planetarium/libplanet/pull/2956
+
 
 Version 0.51.2
 --------------